requirements.txt
Django==2.0.12
django-extensions==2.0.7
djangorestframework==3.9.1
Pillow==7.1.2
pytz==2018.4
six==1.11.0

